#2131: refactored EnumProperty tests into same-package test
