+++ cli/src/main/java/com/devonfw/tools/ide/property/EnumProperty.java
@@ -49,7 +49,7 @@ public class EnumProperty<V extends Enum<V>> extends Property<V> {
   }
 
   @Override
-  public void completeValue(String arg, IdeContext context, Commandlet commandlet, CompletionCandidateCollector collector) {
+  protected void completeValue(String arg, IdeContext context, Commandlet commandlet, CompletionCandidateCollector collector) {
 
     for (V enumConstant : this.valueType.getEnumConstants()) {
       String name = enumConstant.name().toLowerCase(Locale.ROOT);
