+++ cli/src/test/java/com/devonfw/tools/ide/property/EnumPropertyTest.java
@@ -4,8 +4,12 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.CsvSource;
 
 import com.devonfw.tools.ide.commandlet.ContextCommandlet;
+import com.devonfw.tools.ide.commandlet.UpgradeCommandlet;
+import com.devonfw.tools.ide.commandlet.UpgradeMode;
 import com.devonfw.tools.ide.completion.CompletionCandidate;
 import com.devonfw.tools.ide.completion.CompletionCandidateCollector;
 import com.devonfw.tools.ide.completion.CompletionCandidateCollectorDefault;
@@ -51,5 +55,38 @@ class EnumPropertyTest {
     enumProp.completeValue(input, context, new ContextCommandlet(), collector);
 
     assertThat(collector.getCandidates().stream().map(CompletionCandidate::text)).containsExactly(expectedCandidates);
+  }
+
+  @ParameterizedTest
+  @CsvSource({ "'',stable unstable snapshot", "u,unstable", "s,stable snapshot", "st,stable" })
+  void testUpgradeModeCompletion(String input, String expected) {
+    // arrange
+    IdeTestContext context = new IdeTestContext();
+    CompletionCandidateCollector collector = new CompletionCandidateCollectorDefault(context);
+    EnumProperty<UpgradeMode> property = new EnumProperty<>("--mode", false, null, UpgradeMode.class);
+
+    // act
+    property.completeValue(input, context, new UpgradeCommandlet(context), collector);
+
+    // assert
+    assertThat(collector.getCandidates().stream().map(CompletionCandidate::text))
+        .containsExactly(expected.split(" "));
+  }
+
+  /**
+   * Test that no completion candidates are suggested for an unknown prefix.
+   */
+  @Test
+  void testUpgradeModeCompletionNoMatch() {
+    // arrange
+    IdeContext context = new IdeTestContext();
+    CompletionCandidateCollector collector = new CompletionCandidateCollectorDefault(context);
+    EnumProperty<UpgradeMode> property = new EnumProperty<>("--mode", false, null, UpgradeMode.class);
+
+    // act
+    property.completeValue("ss", context, new UpgradeCommandlet(context), collector);
+
+    // assert
+    assertThat(collector.getCandidates()).isEmpty();
   }
 }
